pyproject.toml
[project]
name = "core_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastembed>=0.7.1",
    "flask>=3.1.1",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-core>=0.3.66",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.25",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.4.9",
    "logging>=0.4.9.6",
    "nltk>=3.9.1",
    "numpy>=2.3.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.14.3",
    "sentence-transformers>=4.1.0",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
]
